pyproject.toml
[project]
name = "projet-9"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "faiss-cpu>=1.13.2",
    "fastapi>=0.129.0",
    "httpx>=0.28.1",
    "langchain>=1.2.10",
    "langchain-community>=0.4.1",
    "langchain-mistralai>=1.1.1",
    "mistralai>=1.12.2",
    "openai>=2.24.0",
    "pandas>=3.0.0",
    "python-dotenv>=1.2.1",
    "ragas>=0.4.3",
    "uvicorn>=0.41.0",
]

[dependency-groups]
dev = [
    "pytest>=9.0.2",
    "pytest-cov>=7.0.0",
    "ruff>=0.15.4",
]

[tool.ruff]
extend-exclude = ["notebook/**"]

[tool.ruff.format]
exclude = ["notebook/**"]